+++ relay-server/src/processing/transactions/types/expanded.rs
@@ -10,7 +10,7 @@ use smallvec::smallvec;
 use crate::Envelope;
 use crate::envelope::{ContentType, EnvelopeHeaders, Item, ItemType, Items};
 use crate::managed::{Counted, Managed, Quantities, Rejected};
-use crate::metrics_extraction::transactions::ExtractedMetrics;
+use crate::metrics_extraction::ExtractedMetrics;
 use crate::processing::spans::{Indexed, TotalAndIndexed};
 use crate::processing::transactions::Error;
 use crate::processing::transactions::process::split_indexed_and_total;